Refacto sidebar
Changes to be committed:
	deleted:    components/TaskSidebar.vue
	new file:   components/sidebar/TaskForm.vue
	new file:   components/sidebar/TaskSidebar.vue
	modified:   components/task_item/TaskItem.vue
	new file:   composables/useCategories.composable.ts
	deleted:    composables/useCategoryColor.composable.ts
	new file:   composables/useTaskForm.composable.ts
	modified:   pages/index.vue
